Move get_expt_description and descriptions to more relevant location
