+++ allensdk/brain_observatory/behavior/behavior_ophys_api/behavior_ophys_nwb_api.py
@@ -12,14 +12,18 @@ import xarray as xr
 from pandas.util.testing import assert_frame_equal
 from pynwb import NWBHDF5IO, NWBFile
 
-import allensdk.brain_observatory.roi_masks as roi
 import allensdk.brain_observatory.nwb as nwb
-from allensdk.brain_observatory.behavior.behavior_ophys_api import \
+from allensdk.brain_observatory.behavior.metadata_processing import (
+    get_expt_description
+)
+from allensdk.brain_observatory.behavior.behavior_ophys_api import (
     BehaviorOphysApiBase
+)
 from allensdk.brain_observatory.behavior.schemas import (
     BehaviorTaskParametersSchema, OphysBehaviorMetadataSchema)
-from allensdk.brain_observatory.behavior.trials_processing import \
+from allensdk.brain_observatory.behavior.trials_processing import (
     TRIAL_COLUMN_DESCRIPTION_DICT
+)
 from allensdk.brain_observatory.nwb.metadata import load_pynwb_extension
 from allensdk.brain_observatory.nwb.nwb_api import NwbApi
 from allensdk.brain_observatory.nwb.nwb_utils import set_omitted_stop_time
@@ -27,73 +31,6 @@ from allensdk.core.lazy_property import LazyProperty
 
 load_pynwb_extension(OphysBehaviorMetadataSchema, 'ndx-aibs-behavior-ophys')
 load_pynwb_extension(BehaviorTaskParametersSchema, 'ndx-aibs-behavior-ophys')
-
-
-def _get_expt_description(session_type: str) -> str:
-    """Determine a behavior ophys session's experiment description based on
-    session type.
-
-    Parameters
-    ----------
-    session_type : str
-        A session description string (e.g. OPHYS_1_images_B )
-
-    Returns
-    -------
-    str
-        A description of the experiment based on the session_type.
-
-    Raises
-    ------
-    RuntimeError
-        Behavior ophys sessions should only have 6 different session types.
-        Unknown session types (or malformed session_type strings) will raise
-        an error.
-    """
-    # Experiment descriptions for different session types:
-    # OPHYS_1 -> OPHYS_6
-    ophys_1_3 = dict.fromkeys(
-        ["OPHYS_1", "OPHYS_3"],
-        ("2-photon calcium imaging in the visual cortex of the mouse "
-         "brain as the mouse performs a visual change detection task "
-         "with a set of natural scenes upon which it has previously been "
-         "trained.")
-    )
-    ophys_4_6 = dict.fromkeys(
-        ["OPHYS_4", "OPHYS_6"],
-        ("2-photon calcium imaging in the visual cortex of the mouse "
-         "brain as the mouse performs a visual change detection task "
-         "with a set of natural scenes that are unique from those on "
-         "which it was previously trained.")
-    )
-    ophys_2_5 = {
-        "OPHYS_2": ("2-photon calcium imaging in the visual cortex of the "
-                    "mouse brain as the mouse is shown images from a "
-                    "change detection task with a set of natural scenes "
-                    "upon which it has previously been trained, but with "
-                    "the lick-response sensor withdrawn (passive/open "
-                    "loop mode)."),
-        "OPHYS_5": ("2-photon calcium imaging in the visual cortex of the "
-                    "mouse brain as the mouse is shown images from a "
-                    "change detection task with a set of natural scenes "
-                    "that are unique from those on which it was "
-                    "previously trained, but with the lick-response "
-                    "sensor withdrawn (passive/open loop mode)."),
-    }
-    expt_description_dict = {**ophys_1_3, **ophys_2_5, **ophys_4_6}
-
-    # Session type string will look something like: OPHYS_4_images_A
-    truncated_session_type = "_".join(session_type.split("_")[:2])
-
-    try:
-        return expt_description_dict[truncated_session_type]
-    except KeyError as e:
-        e_msg = (
-            f"Encountered an unknown session type "
-            f"({truncated_session_type}) when trying to determine "
-            f"experiment descriptions. Valid session types are: "
-            f"{expt_description_dict.keys()}")
-        raise RuntimeError(e_msg) from e
 
 
 class BehaviorOphysNwbApi(NwbApi, BehaviorOphysApiBase):
@@ -114,7 +51,7 @@ class BehaviorOphysNwbApi(NwbApi, BehaviorOphysApiBase):
             institution="Allen Institute for Brain Science",
             keywords=["2-photon", "calcium imaging", "visual cortex",
                       "behavior", "task"],
-            experiment_description=_get_expt_description(session_type)
+            experiment_description=get_expt_description(session_type)
         )
 
         # Add stimulus_timestamps to NWB in-memory object:
